requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
discord
python-dotenv
torch==1.11.0+cpu
torchvision==0.12.0+cpu
detoxify
pillow
